newport: address corruption; fix / add missing FIFO waits; comment on what we should do!
I was seeing corruption on my Newport XL8 on an R4600PC-100.
The R4600 can burst uncached writes much faster than an R4000
as it buffers them in a little FIFO so the pipeline doesn't
immediately stall until it's done.

Anyway;

* add a couple of missing FIFO waits
* add TODO comments around how we SHOULD be doing the FIFO waits rather than
  how it's currently being done.
* add TODO comments where the go.hostrw0 / go.zpattern register is used
  directly by XAA, where we don't get to monitor the GFIFO and wait
  appropriately.

I think to be technically correct the alternate XAA methods for those two
direct register uses should be used - where a driver allocated temporary memory
buffer is populated instead, and then it's up to us to push it into a NIC.

Tested:

* R4600PC-100, XL24, 8 bit mode
* R4600PC-100, XL24, 24 bit mode
* R4600PC-100, XL8, 8 bit mode