newport: update some documentation around the colour plane / WRMASK stuff
I think the writemask stuff is a bit "special" and un-needed here.
In any case, it's a function of the pixel write format into the
framebuffer, NOT the HOSTRW input format.

(ie, we can configure DRAWMODE1 to accept packed RGB888 / RGB444
data but write out 8 bit RGB332, so with a little bit of refactoring
we should be able to do naughty stuff..)