newport: add some more debugging around acceleration register IO
This isn't the register IO itself, but is the shadow register calls.
But it's useful for debugging.

It turns out I've elicited a fun corner case - the initial values
aren't correct, so that red flash is because it's trying to write
COLORI=0, but the shadow value has 0 and the REAL register value
is something else!  I'll have to clean that up in a subsequent
commit.