newport: register a shadowfb callback when accel is enabled; add debugging
As said in the code, it looks like there's a bunch of X server primitives
which are not being accelerated and instead are ending up in the shadow
framebuffer.  Unfortunately for us this breaks on newport because we don't
have cheap access to the underlying framebuffer.

So at least add a callback and document what's going on so I can dig into
it all a little further.